+++ tools/_release.sh
@@ -36,6 +36,7 @@ if [ "$RELEASE_GPG_KEY" = "" ]; then
         BF6BCFC89E90747B9A680FD7B6029E8500F7DB16
         86379B4F0AF371B50CD9E5FF3402831161D1D280
         20F201346BF8F3F455A73F9A780CC99432A28621
+        F2871B4152AE13C49519111F447BF683AA3B26C3
     "
     for key in $TRUSTED_KEYS; do
         if gpg2 --with-colons --card-status | grep -q "$key"; then
@@ -157,29 +158,25 @@ done
 mkdir "dist.$version"
 for pkg_dir in $SUBPKGS
 do
-  mv $pkg_dir/dist "dist.$version/$pkg_dir/"
+  mv $pkg_dir/dist/* "dist.$version"
 done
 
 echo "Testing packages"
 cd "dist.$version"
-# start local PyPI
-python -m http.server $PORT &
 # cd .. is NOT done on purpose: we make sure that all subpackages are
-# installed from local PyPI rather than current directory (repo root)
+# installed from local archives rather than current directory (repo root)
 VIRTUALENV_NO_DOWNLOAD=1 virtualenv ../venv
 . ../venv/bin/activate
 pip install -U setuptools
 pip install -U pip
-# Now, use our local PyPI. Disable cache so we get the correct KGS even if we
-# (or our dependencies) have conditional dependencies implemented with if
-# statements in setup.py and we have cached wheels lying around that would
-# cause those ifs to not be evaluated.
+# Now, use our local archives. Disable cache so we get the correct packages even if
+# we (or our dependencies) have conditional dependencies implemented with if
+# statements in setup.py and we have cached wheels lying around that would cause
+# those ifs to not be evaluated.
 python ../tools/pip_install.py \
   --no-cache-dir \
-  --extra-index-url http://localhost:$PORT \
+  --find-links . \
   $SUBPKGS
-# stop local PyPI
-kill $!
 cd ~-
 
 # get a snapshot of the CLI help for the docs